Add files via upload
Consider a path optimization problem with resources and locks. We are given a weighted
undirected graph, and the goal is to travel as cheaply as possible from a given start vertex to a
given goal vertex. Here the problem is that some of the vertices are blocked, and can only be
unblocked if the agent is carrying an appropriate resource (key), which can be found in other
vertices of the graph (which may also be blocked or not). A key is intended to act on a lock
with the same ID. An agent at a vertex automatically picks up all the keys at that vertex just
before starting the next move. Entering a locked vertex is only possible when the agent has the
appropriate key, at which point the lock and key both disappear. An agent can do a
traverse action from one node to another whose cost is equal to the edge weight connecting the
two nodes. The action always succeeds if the target vertex is unblocked (or becomes unblocked
due to having the correct key(s)), and fails otherwise. For a failed action, the agent stays at the
same location but still pays the cost of traversal. Implement an agent using A* search that picks
the move with best immediate heuristic value.